Go Modules and refactor file layout
