+++ cmd/receipt-sink-service/Dockerfile
