+++ cmd/sdx-store-service/Dockerfile
