+++ cmd/sdx-survey-gateway-service/Dockerfile
