+++ cmd/sdx-survey-gateway-service/docker-compose.yml
