+++ cmd/sdx-survey-gateway-service/health.go
