+++ go.mod
@@ -1,0 +1,9 @@
+module github.com/ONSdigital/sdx-evolution
+
+go 1.13
+
+require (
+	github.com/garyburd/redigo v1.6.0
+	github.com/gorilla/mux v1.7.3
+	github.com/streadway/amqp v0.0.0-20200108173154-1c71cc93ed71
+)
