+++ internal/rabbit/rabbit.go
